% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanGroupData.r
\name{addPeopleLocationsRelations}
\alias{addPeopleLocationsRelations}
\title{Inside function of \code{\link[=cleanGroupData]{cleanGroupData()}}}
\usage{
addPeopleLocationsRelations(x, relations)
}
\arguments{
\item{x}{A subset of the relation table providing
where a single group is seen}

\item{relations}{Relations Veritas table}
}
\value{
New rows for the relation table
}
\description{
Find people <-> locations relations not present in relation table
}
