% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcheck.R
\name{analyseBatch}
\alias{analyseBatch}
\title{Visualise the batch run}
\usage{
analyseBatch(batchobj, draw = c("static", "interactive"), maxwidth = 5, ...)
}
\arguments{
\item{batchobj}{An object generated by \link{batchRun}}

\item{draw}{A character indicating whether one wants to make a static or an interactive igraph plot}

\item{maxwidth}{The maximum width of the edges}

\item{...}{Additional arguments that are being passed to igraphs plotting functions.}
}
\description{
Draws the decision net with thickness of edges corresponding to the number of scripts that went through that edge.
Final nodes (END nodes) display the number of submissions in them.
}
