% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSocialNetwork.r
\name{createSingleSocialNetwork}
\alias{createSingleSocialNetwork}
\title{Social network constructor}
\usage{
createSingleSocialNetwork(veritas_pid, group_scale = TRUE)
}
\arguments{
\item{veritas_pid}{Veritas data for a single participant.
See \code{\link[=transformData]{transformData()}}.}

\item{group_scale}{Boolean. Declare if the number of vertex generated
from group are scaled to the square root.}
}
\value{
A single igraph network
}
\description{
This function create a social network for a single participant,
}
\details{
Generate an igraph networks which include the participant and
the people linked to him (either reported individually of from groups),
}
