% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSocialNetwork.r
\name{createSocialNetworks}
\alias{createSocialNetworks}
\title{Constructor of social networks}
\usage{
createSocialNetworks(veritas_split, group_scale = TRUE)
}
\arguments{
\item{veritas_split}{A list of veritas data for a all participants.
See \code{\link[=transformData]{transformData()}}.}

\item{group_scale}{Boolean. Declare if the number of vertex generated
from group are scaled to the square root.}
}
\value{
A List of igraph networks.
}
\description{
This function return the set of all social networks from a Veritas dataset.
}
\details{
Generates a set of igraph networks,
each of which includes a participant and the people linked to him
(either reported individually or from groups).
}
\seealso{
This function is a wrapper of \code{\link[=createSingleSocialNetwork]{createSingleSocialNetwork()}}.
}
