% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{did_call}
\alias{did_call}
\title{Check whether a function was called by code}
\usage{
did_call(fun, pkg, obj, args = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{fun}{The function that one wants to track, provided as a character}

\item{pkg}{The packages that the function comes from, provided as a character}

\item{obj}{An object passed on by the evaluator (need not be supplied by the user)}

\item{args}{A list of single value arguments that need to have been passed to the tracked function, for did_call to return true}

\item{tol}{The tolerance for the arguments}
}
\value{
true or false
}
\description{
Function for checking whether submitted code calls a specific function.
Only tracks functions that are called with the package attached (i.e. can track plot, but not base::plot).
Cannot track functions that load datasets (i.e. read.csv).
}
\seealso{
\link{runSingle} for an example
}
