% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-net.R
\name{drawNet}
\alias{drawNet}
\title{Draw a decision network}
\usage{
drawNet(net, draw = c("none", "interactive", "static"), ...)
}
\arguments{
\item{net}{A list of nodes (see \link{node}). No node should be called END. One node should be called start (the first)}

\item{draw}{Whether a static or an interactive plot (in the letter the nodes can be dragged around)}

\item{...}{Additional parameters to be passed on to the igraph plotting functions}
}
\description{
Draws a decision network for visual inspection
}
