% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcheck.R
\name{extractFeedback}
\alias{extractFeedback}
\title{Obtain feedback
Function for turning a batch check object into a dataframe that contains the feedback for each student.}
\usage{
extractFeedback(batchobj)
}
\arguments{
\item{batchobj}{An object generated by \link{batchRun}}
}
\value{
A dataframe with two columns. 1. The filenames of the R scripts, 2. The output that was generated by VeRitas
}
\description{
Obtain feedback
Function for turning a batch check object into a dataframe that contains the feedback for each student.
}
