% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcheck.R
\name{extractScripts}
\alias{extractScripts}
\title{Function for getting all scripts from specific nodes}
\usage{
extractScripts(batchobj, node, target)
}
\arguments{
\item{batchobj}{An object generated by \link{batchRun}}

\item{node}{A character indicating the node that one wants to obtain the scripts from}

\item{target}{(optional) If one wants to extract scripts that have moved between two specific nodes, also include a target node.}
}
\description{
Extract script names from a given node.
}
