% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{Groups data}
\format{
A data frame with 68 rows and 4 variables:
\describe{
\item{pid}{participants unique identifiers}
\item{group_id}{group unique identifiers}
\item{group_name}{group names}
\item{group_size}{Number of people in the group}
}
}
\usage{
groups
}
\description{
A dumy dataset containing the reported groups of 50 participants
}
\keyword{datasets}
