% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{has_each_val}
\alias{has_each_val}
\title{Check whether each value in a vector is produced by code}
\usage{
has_each_val(
  values,
  obj,
  text = FALSE,
  tol = sqrt(.Machine$double.eps),
  absval = FALSE
)
}
\arguments{
\item{obj}{An object passed on by the evaluator (need not be supplied by the user)}

\item{text}{A boolean that indicates whether also the code itself should be scanned for the values (and not just the output/generated variables)}

\item{tol}{The tolerance}

\item{absval}{Whether absolute values should be matched. If true, values in obj and vec are first turned into their absolute values before checking.}

\item{value}{The vector of values that one is looking for}
}
\value{
true or false
}
\description{
Function for checking whether submitted code produces each value of a certain vector.
The search does not care where the values are found. Also, it does not account for multiples
(i.e. it does not distinguish between a code that produces a certain value once or one that produces it twice)
}
