% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finders.R
\name{has_vec}
\alias{has_vec}
\title{Check whether a vector is produced by code}
\usage{
has_vec(vec, obj, ord = TRUE, tol = sqrt(.Machine$double.eps), absval = FALSE)
}
\arguments{
\item{vec}{The vector that one is looking for}

\item{obj}{An object passed on by the evaluator (need not be supplied by the user)}

\item{ord}{A boolean that indicates whether the vector should be ordered}

\item{tol}{The tolerance}

\item{absval}{Whether absolute values should be matched. If true, values in obj and vec are first turned into their absolute values before checking.}
}
\value{
true or false depending on whether the submitted code produced the vector, either by storing it into a variable, or by putting it on the screen.
}
\description{
Function for checking whether submitted code contains a certain vector.
The search is performed recursively, including columns of data frames.
}
