% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactSnMeasures.r
\name{interactSnMeasures}
\alias{interactSnMeasures}
\title{Calculate SN measures from InterAct data}
\usage{
interactSnMeasures(veritas_data)
}
\arguments{
\item{veritas_data}{Transformed data generated from \code{\link[=transformData]{transformData()}}}
}
\value{
Table of social network measures. See vignette (CREATE VIGNETTE).
}
\description{
Calculate social network measures for Veritas data collected from
the InterAct research project.
See \href{https://teaminteract.ca}{InterAct web site} for more information.
}
