% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{Locations data}
\format{
A data frame with 1140 rows and 4 variables:
\describe{
\item{pid}{participants unique identifiers}
\item{location_id}{location unique identifiers}
\item{location_category}{location categories}
\item{location_name}{location names}
}
}
\usage{
locations
}
\description{
A dumy dataset containing the reported locations of 50 participants
}
\keyword{datasets}
