% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-net.R
\name{makeTrackEnv}
\alias{makeTrackEnv}
\title{function for making an environment that redefines all test functions,
so did_call can be evaluated}
\usage{
makeTrackEnv()
}
\description{
function for making an environment that redefines all test functions,
so did_call can be evaluated
}
