% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people}
\alias{people}
\title{People data}
\format{
A data frame with 190 rows and 4 variables:
\describe{
\item{pid}{participants unique identifiers}
\item{people_id}{people unique identifiers}
\item{people_firstname}{First names}
\item{people_relation}{relation categories}
}
}
\usage{
people
}
\description{
A dumy dataset containing the reported people of 50 participants
}
\keyword{datasets}
