% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.veritas.social.network.r
\name{plot.veritas.social.network}
\alias{plot.veritas.social.network}
\title{Plot social network}
\usage{
\method{plot}{veritas.social.network}(g, groups = TRUE, labels = FALSE)
}
\arguments{
\item{groups}{Boolean. Define if groups are displayed}

\item{labels}{Boolean. Define if labels are displayed}

\item{networks}{An object of class \code{veritas.social.network}}
}
\description{
Plot a social networks generated with \code{\link[=createSocialNetwork]{createSocialNetwork()}}
}
\details{
Method of the \code{\link[=plot]{plot()}} generic function that run on object
of class \code{veritas.social.network}.
}
