% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecheck.R
\name{prepareAssignment}
\alias{prepareAssignment}
\title{Assignment setup}
\usage{
prepareAssignment(net, referencefile)
}
\arguments{
\item{net}{A named list of nodes (see \link{node}) that describes what checks should be performed and what feedback they yield. The first node should be named start, and any final nodes should be named END. No other nodes should start with these words in their name.}

\item{referencefile}{An R script that contains a reference solution. The checks in the nodes use values from the reference solution.}
}
\value{
Returns an initialized assignment, which is in essence a list that contains the net, the functions that should be tracked, and the output of the reference file.
}
\description{
Uses an assessment net (a named list of nodes) and a reference file to prepare an assignment.
}
\seealso{
\link{runSingle} for a simple example.
}
