% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.validation.r
\name{print.partial.validation}
\alias{print.partial.validation}
\title{Print partial validation information}
\usage{
\method{print}{partial.validation}(x, ...)
}
\arguments{
\item{x}{Object generated by \code{\link[=validateData]{validateData()}} for which columns are not valid}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
This is a method of the \code{\link[=print]{print()}} function for object
of class \code{partial.validation} generated from \code{\link[=validateData]{validateData()}}.
}
\details{
Print on the command line interface validation information on
column names only. Other validation could not be compute.
}
