% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.validation.NAs.R
\name{print.validation.NAs}
\alias{print.validation.NAs}
\title{Print NAs validation}
\usage{
\method{print}{validation.NAs}(x, ...)
}
\arguments{
\item{x}{Object returned by \code{\link[=validateNAs]{validateNAs()}}}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
This is a method of the \code{\link[=print]{print()}} function for object
of class \code{validation.NAs} generated from \code{\link[=validateNAs]{validateNAs()}}.
}
\details{
Print on the command line interface the column names with
founded non standard NAs (i.e., values < 0 or == 99).
}
