% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.validation.r
\name{print.validation}
\alias{print.validation}
\title{Print validation information}
\usage{
\method{print}{validation}(x, ...)
}
\arguments{
\item{x}{Object generated by \code{\link[=validateData]{validateData()}}}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
This is a method of the \code{\link[=print]{print()}} function for object
of class \code{validation} generated from \code{\link[=validateData]{validateData()}}.
}
\details{
Print on the command line interface all validation information on:
\itemize{
\item The presence of appropriate column names
\item Duplicated people, groups and locations ids
\item PIDs in people, groups and relations tables not in locations table
\item Nodes in relations tables
\item Group size
}
}
