% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relations}
\alias{relations}
\title{Relations data}
\format{
A data frame with 626 rows and 4 variables:
\describe{
\item{pid}{participants unique identifiers}
\item{node_1}{node 1 unique identifiers}
\item{relation_type}{Types of relations}
\item{node_2}{node 2 unique identifiers}
}
}
\usage{
relations
}
\description{
A dumy dataset containing the reported relations between people,
groups and locations for 50 participants
}
\keyword{datasets}
