% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcheck.R
\name{runBatch}
\alias{runBatch}
\title{Function for running a set of scripts
Use with caution!
If you run this script, all submitted R scripts will be executed automatically.
Be sure that you trust these scripts - they could severly harm your computer.}
\usage{
runBatch(
  dirname,
  assignment,
  pattern = ".*\\\\.R$",
  rec = FALSE,
  showfiles = TRUE
)
}
\description{
Function for running a set of scripts
Use with caution!
If you run this script, all submitted R scripts will be executed automatically.
Be sure that you trust these scripts - they could severly harm your computer.
}
