% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecheck.R
\name{runSingle}
\alias{runSingle}
\title{Evaluate a single script.}
\usage{
runSingle(studentfile, prepared)
}
\arguments{
\item{studentfile}{The name (and location) of the file that is to be evaluated.}

\item{prepared}{A prepared assignment using \link{prepareAssignment}}
}
\value{
The feedback is shown on the screen.
}
\description{
This function evaluates a single .R file based on a prepared assignment.
}
\examples{
# example code
library(VeRitas)
net <- list(start=node(has_val(ref$sol),
                       goT=S2,goF=S1,
                       doT="[✓] Sum is found!",
                       doF="[x] Sum not found"),

            S1=node(has_val(ref$wrongsol),
                    goT=END,goF=END,
                    doT="[?] Did you square?"),

             S2=node(did_call("sum","base"),
                    goT=END,goF=END,
                    doF="[?] Could the sum function help?")
)
MyAssignment <- prepareAssignment(net,"../../examples/01-simplest/ref1.R")
runSingle("../../examples/01-simplest/code2.R",MyAssignment)
}
