% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialNetworkMeasures.r
\name{socialNetworkMeasures}
\alias{socialNetworkMeasures}
\title{Calculate social network measures}
\usage{
socialNetworkMeasures(social_networks)
}
\arguments{
\item{social_networks}{networks generated with \code{\link[=createSocialNetwork]{createSocialNetwork()}}}
}
\value{
Data frame of network measures by participant
}
\description{
This function calculate network measures from social networks
generated from Veritas data.
}
\details{
Calculate the following measures:
\itemize{
\item \emph{people_degree:} Number of individually reported people
\item \emph{group_degree:} Number of people generated from groups
\item \emph{degree:} Number of people
\item \emph{clustering:} Participant clustering coefficient. Equivalent of the
network density without relations between the participants and the alters.
\item \emph{simmelian:} Simmelian brokerage.
See \href{https://doi.org/10.1007/s10955-013-0722-z}{Latora et al., 2013}
}
}
