% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.validation.r
\name{summary.partial.validation}
\alias{summary.partial.validation}
\title{Summarize partial validation information}
\usage{
\method{summary}{partial.validation}(object, ...)
}
\arguments{
\item{object}{Object of class \code{partial.validation}
generated by \code{\link[=validateData]{validateData()}}}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
This function report that column names fail, and the other validation
are unknown since they could not be test in absence of proper
column names
}
\details{
This is a method of the \code{\link[=summary]{summary()}} generic function.
Print on the Command Line Interface pass or unknown for each validation
test
}
