% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.validation.r
\name{summary.validation}
\alias{summary.validation}
\title{Summarize validation information}
\usage{
\method{summary}{validation}(object, ...)
}
\arguments{
\item{object}{Object of class \code{validation} generated by \code{\link[=validateData]{validateData()}}}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
This function report if each validation test pass of fail.
}
\details{
This is a method of the \code{\link[=summary]{summary()}} generic function.
Print on the Command Line Interface pass or fail for each validation test
}
