% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.veritas.r
\name{summary.veritas.pid}
\alias{summary.veritas.pid}
\title{Summarize veritas data information for a single participant.}
\usage{
\method{summary}{veritas.pid}(object, ...)
}
\arguments{
\item{object}{An Object of class \code{veritas.pid}, which
is an element of the list generated by \code{\link[=transformData]{transformData()}}.}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
Print the number of locations, people, groups and relations
of a single participant
}
