% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.veritas.r
\name{summary.veritas.split}
\alias{summary.veritas.split}
\title{Summarize veritas data information}
\usage{
\method{summary}{veritas.split}(object, ...)
}
\arguments{
\item{object}{Object of class \code{veritas.split} generate by \code{\link[=transformData]{transformData()}}}

\item{...}{Only used to remain consistent with plot generic function}
}
\description{
Print the number of participants, locations, people, groups and relations
in the whole veritas data base
}
\details{
This is method of the \code{\link[=summary]{summary()}} generic function.
Print on the Command Line Interface summary information.
}
