% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.r
\name{transformData}
\alias{transformData}
\title{Split veritas data by pid}
\usage{
transformData(locations, people, groups, relations)
}
\arguments{
\item{locations}{Location Veritas table}

\item{people}{People Veritas table}

\item{groups}{Groups Veritas table}

\item{relations}{Relations Veritas table}
}
\value{
A list of list of tables by participant.
}
\description{
This function seperate the four input tables by participant.
}
\details{
For every participant, it extract their locations, people, groups and
relations, and create NULL object if the participant have none of them
(e.g., return NULL for people if the participant have not reported any).
}
\seealso{
Check vignette (ADD VIGNETTE) for the format of the input tables.
}
