% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateData.R
\name{validateData}
\alias{validateData}
\title{Validate Veritas input tables}
\usage{
validateData(locations, people, groups, relations)
}
\arguments{
\item{locations}{Location Veritas table}

\item{people}{People Veritas table}

\item{groups}{Groups Veritas table}

\item{relations}{Relations Veritas table}
}
\value{
An object of class \code{validation} which is a
list of validation information.
}
\description{
This function check if the input tables have the proper format to run
\code{\link[=transformData]{transformData()}}.
}
\details{
Fun by \code{\link[=transformData]{transformData()}} to test if the input data can run by the function.
}
\seealso{
See vignette (ADD VIGNETTE!!!) for more information
on the required format.
}
