% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateNAs.R
\name{validateNAs}
\alias{validateNAs}
\title{Validate NAs in numerical variables}
\usage{
validateNAs(locations, people, groups, relations)
}
\arguments{
\item{locations}{Locations Veritas table}

\item{people}{People Veritas table}

\item{groups}{Groups Veritas table}

\item{relations}{Relations Veritas table}
}
\value{
An object of class \code{validation.NAs}, which is a
list comprising \code{valide}, a boolean indicating wether there potentialy
is non standard NAs, and \code{details}, the number of non standard NAs
by variables in each table.
}
\description{
Check for negative vlaues or 99 in numerical variables
}
\details{
This step was added as non standard NAs were found in Veritas data
of the InterAct project. It is run by \code{\link[=transformData]{transformData()}}, but only
generate a \code{warning} if non standard NAs are found.
}
