% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runEval}
\alias{runEval}
\title{Main function for running a single R script}
\usage{
runEval(
  Rfile,
  totrack,
  ref,
  net,
  csv = "",
  killFuns = c("`?`", "setwd", "system", "file.choose", "install.packages", "View"),
  verbose = FALSE
)
}
\description{
Main function for running a single R script
}
